# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Labzz Chat API",
        "description": "Realtime chat API built with Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/conversations": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "summary": "List user conversations",
                "operationId": "listConversations",
                "responses": {
                    "200": {
                        "description": "List of conversations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "private"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2026-03-16 10:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Create conversation",
                "operationId": "createConversation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "participants"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "private"
                                    },
                                    "participants": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conversation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "private"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2026-03-16 10:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send message",
                "operationId": "1aa21572307a1c2e85281f5e04599031",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "conversation_id",
                                    "content"
                                ],
                                "properties": {
                                    "conversation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Hello world"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/conversations/{id}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List conversation messages",
                "operationId": "9700f8fe888769c278fe6ab2ac8cce26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "a5385ffca73f113abef00ea604d1ca58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/auth/2fa/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify two factor authentication",
                "operationId": "3f6c7f84605529d37a5827d5d76aded5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_token",
                                    "code"
                                ],
                                "properties": {
                                    "login_token": {
                                        "type": "string",
                                        "example": "uuid-token"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Conversations",
            "description": "Conversations"
        },
        {
            "name": "Messages",
            "description": "Messages"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}